crawler/requirements.txt
playwright==1.38.0
beautifulsoup4
httpx
tqdm
python-dotenv
